scripts/training/requirements.txt
tensorflow-cpu==2.10
tensorflow-directml-plugin
tf2onnx
pillow
numpy
